requirements.txt
Flask
natto-py
numpy
scikit-learn
scipy